# Make script processor Run method concurrency safe